requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
scikit-learn
PyYAML
mlflow
dagshub
fastapi
uvicorn
# -e.